create poll with match and poll status

+++ app/rest_api/api/club/club.py
@@ -2,16 +2,18 @@ from typing import Annotated
 
 from fastapi import APIRouter, Depends, Query
 from fastapi_filter import FilterDepends
-from sqlalchemy import or_
+from sqlalchemy import and_, delete, exists, or_
 from sqlalchemy.orm import Session, joinedload
 
 from app.core.deps import get_db
 from app.core.token import get_current_user
+from app.helper.exception import ClubNotFoundException, JoinClubNotFoundException
 from app.model.club import Club, JoinClub
 from app.model.match import Match
 from app.model.position import JoinPosition
 from app.model.profile import Profile
 from app.rest_api.schema.club.club import (
+    ClubResponseSchema,
     ClubSchema,
     FilterClubSchema,
     UpdateClubSchema,
@@ -41,14 +43,16 @@ def create_club(
     db.add(club)
     db.commit()
 
-    join_club = JoinClub(clubs_seq=club.seq, user_seq=token.seq, role="회장")
+    join_club = JoinClub(
+        clubs_seq=club.seq, user_seq=token.seq, role="회장", accepted=True
+    )
     db.add(join_club)
     db.commit()
 
     return {"success": True}
 
 
-@club_router.get("/{club_seq}")
+@club_router.get("/{club_seq}", response_model=ClubResponseSchema)
 def get_club(
     token: Annotated[str, Depends(get_current_user)],
     club_seq: int,
@@ -56,8 +60,8 @@ def get_club(
 ):
     club = db.query(Club).filter(Club.seq == club_seq).first()
 
-    # if club is None:
-    #     raise ClubNotFoundException
+    if club is None:
+        raise ClubNotFoundException
 
     return club
 
@@ -71,8 +75,8 @@ def update_club(
 ):
     club = db.query(Club).filter(Club.seq == club_seq).first()
 
-    # if not club:
-    #     raise ClubNotFoundException
+    if club is None:
+        raise ClubNotFoundException
 
     for key, value in update_club_data.dict(exclude_none=True).items():
         setattr(club, key, value)
@@ -82,7 +86,7 @@ def update_club(
     return {"success": True}
 
 
-@club_router.get("")
+@club_router.get("", response_model=list[ClubResponseSchema])
 def filter_clubs(
     token: Annotated[str, Depends(get_current_user)],
     club_filter: FilterClubSchema = FilterDepends(FilterClubSchema),
@@ -90,33 +94,80 @@ def filter_clubs(
     per_page: int = Query(10, title="페이지당 수", ge=1, le=100),
     db: Session = Depends(get_db),
 ):
-    query = db.query(
-        Club.seq,
-        Club.name,
-        Club.location,
-        Club.age_group,
-        Club.skill,
-        Club.membership_fee,
-        Club.emblem_img,
-    )
+    query = db.query(Club)
     query = club_filter.filter(query)
     offset = (page - 1) * per_page
     query = query.limit(per_page).offset(offset)
-
-    clubs = [
-        {
-            "seq": seq,
-            "name": name,
-            "location": location,
-            "age_group": age_group,
-            "skill": skill,
-            "membership_fee": membership_fee,
-            "emblem_img": emblem_img,
-        }
-        for seq, name, location, age_group, skill, membership_fee, emblem_img in query.all()
-    ]
+    clubs = query.all()
 
     return clubs
+
+
+@club_router.post("/{club_seq}/join")
+def join_club(
+    club_seq: int,
+    token: Annotated[str, Depends(get_current_user)],
+    db: Session = Depends(get_db),
+):
+    join_status = db.query(
+        exists().where(
+            (JoinClub.user_seq == token.seq) & (JoinClub.clubs_seq == club_seq)
+        )
+    ).scalar()
+
+    if not join_status:
+        join_club = JoinClub(user_seq=token.seq, clubs_seq=club_seq, role="회원")
+        db.merge(join_club)
+        db.commit()
+        db.flush()
+
+    return {"success": True}
+
+
+@club_router.patch("/{club_seq}/accept")
+def accept_club(
+    club_seq: int,
+    user_seq: int,
+    token: Annotated[str, Depends(get_current_user)],
+    db: Session = Depends(get_db),
+):
+    club = db.query(Club).filter(Club.seq == club_seq).first()
+    if not club:
+        raise ClubNotFoundException
+
+    join_club = (
+        db.query(JoinClub)
+        .filter(
+            JoinClub.club_seq == club_seq,
+            JoinClub.user_seq == user_seq,
+        )
+        .one()
+    )
+
+    if not join_club:
+        raise JoinClubNotFoundException
+
+    # TODO: validate club owner / matcher poster
+
+    join_club.accepted = True
+
+    return {"success": True}
+
+
+@club_router.delete("/{club_seq}/quit")
+def quit_club(
+    club_seq: int,
+    token: Annotated[str, Depends(get_current_user)],
+    db: Session = Depends(get_db),
+):
+    join_club = delete(JoinClub).where(
+        and_(JoinClub.user_seq == token.seq, JoinClub.clubs_seq == club_seq)
+    )
+    db.execute(join_club)
+    db.commit()
+    db.flush()
+
+    return {"success": True}
 
 
 @club_router.delete("/{club_seq}")
@@ -127,8 +178,8 @@ def delete_club(
 ):
     club = db.query(Club).filter(Club.seq == club_seq).first()
 
-    # if not club:
-    #     raise ClubNotFoundException
+    if club is None:
+        raise ClubNotFoundException
 
     db.delete(club)
     db.commit()
@@ -148,7 +199,7 @@ def get_members(
     members = (
         db.query(Profile)
         .join(JoinClub, Profile.user_seq == JoinClub.user_seq)
-        .filter(JoinClub.clubs_seq == club_seq)
+        .filter(JoinClub.clubs_seq == club_seq, JoinClub.accepted == True)
         .options(joinedload(Profile.join_position).joinedload(JoinPosition.position))
         .all()
     )
